refactor:  change the file structure and name space
