+++ ParkingLot/ErrorHandling/OutOfCapacityException.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Runtime.Serialization;
 
-namespace ParkingLotProj
+namespace ParkingLotProj.ErrorHandling
 {
     [Serializable]
     public class OutOfCapacityException : Exception
